create a class to handle the output just so things are a bit cleaner
