--- packages/bundler-plugin-core/src/utils/buildEndHelper.ts
@@ -1,5 +1,0 @@
-import { type Output } from "../types";
-
-export const buildEndHelper = (output: Output) => {
-  output.duration = Date.now() - (output.builtAt ?? 0);
-};
